refactor(blazwitcher-extension): :recycle: restructure plugins module
