+++ .changeset/angry-planets-live.md
@@ -1,0 +1,6 @@
+---
+"blazwitcher": minor
+---
+
+feat: refactor plugins module structure and add keyboard navigation
+feat: 重构plugins模块结构并添加键盘导航功能
