+++ packages/blazwitcher-extension/options/index.tsx
@@ -1,5 +1,5 @@
 import './options.css'
-import { SettingPanels } from '~plugins/settingPanels'
+import { SettingPanels } from '~plugins/ui/setting-panels'
 import { useTheme } from '~sidepanel/hooks/useTheme'
 
 // 自定义设置页面
