+++ packages/blazwitcher-extension/plugins/commands/index.ts
@@ -1,0 +1,2 @@
+export { filterByBookmarkPlugin, filterByHistoryPlugin, filterByTabPlugin } from './filters'
+export { settingPlugin } from './setting'
