+++ packages/blazwitcher-extension/plugins/ui/setting-panels/appearance.tsx
