+++ packages/blazwitcher-extension/plugins/ui/setting-panels/changelog.tsx
