+++ packages/blazwitcher-extension/plugins/ui/setting-panels/keyboard.tsx
