+++ packages/blazwitcher-extension/plugins/ui/setting-panels/search.tsx
