+++ packages/blazwitcher-extension/sidepanel/footer/active-item-description.tsx
@@ -2,7 +2,7 @@ import EnterSvg from 'react:~assets/enter.svg'
 import { useAtomValue } from 'jotai'
 import { useCallback, useMemo } from 'react'
 import styled from 'styled-components'
-import { usePluginClickItem } from '~plugins/render-item'
+import { usePluginClickItem } from '~plugins/ui/render-item'
 import { FOOTER_DESCRIPTION_I18N_MAP } from '~shared/constants'
 import { ItemType } from '~shared/types'
 import { getItemType, handleItemClick, isPluginItem } from '~shared/utils'
