+++ packages/blazwitcher-extension/sidepanel/index.tsx
@@ -9,8 +9,8 @@ import { MAIN_CONTENT_CLASS } from '~shared/constants'
 import { handleItemClick, orderList, searchWithList, splitToGroup } from '~shared/utils'
 
 import plugins from '~plugins'
-import { matchPlugin } from '~plugins/helper'
-import { RenderPluginItem, usePluginClickItem } from '~plugins/render-item'
+import { matchPlugin } from '~plugins'
+import { RenderPluginItem, usePluginClickItem } from '~plugins/ui/render-item'
 import { ItemType, type ListItemType } from '~shared/types'
 import { i18nAtom, searchConfigAtom } from '~sidepanel/atom'
 import useOriginalList from '~sidepanel/hooks/useOriginalList'
